backend/requirements.txt
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.7
supabase==1.2.0

# AI/ML Libraries
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
huggingface-hub==0.17.3
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# API Integration
requests==2.31.0
httpx==0.25.0
aiohttp==3.8.6

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.4.2
marshmallow==3.20.1

# Government APIs
zeep==4.2.1
xmltodict==0.13.0

# Health Data Standards
fhir.resources==7.0.2
hl7apy==1.3.4

# Monitoring
prometheus-client==0.17.1